refactor(textobject): move code to separate module
